Create LiveImageStore classes
Create the following LiveImageStore classes

DirectoryStructureUpdater class which acts as a proxy to
the LiveImageStore clients. I am not yet convinced that this
class is really necessary but will keep it for the moment as
we can easily refactor it out later.

Regarding clients

Dummy client class for testing.

LocalFileSystem class which will update the local file system.
This class allows me to think about how to create
LiveImageStore clients that is easy to experiment with before
using the more complicated Amazon one. It also serves as a
valid client in its own right, some people might not want to
use Amazon.